Rename HelloAndroidCamera2 to HelloAndroid
Now that the old HelloAndroid app is gone, this is the only Android
example app in the repo, so it takes the shorter name. Also moves
android_smoke_test.sh into the app's own folder, since it's no longer
shared between two apps.
