+++ apps/CMakeLists.txt
@@ -55,7 +55,7 @@ add_app(fft)
 add_app(gaussian_blur)
 add_app(hannk)
 add_app(harris)
-# add_app(HelloAndroidCamera2)  # don't build HelloAndroidCamera2 here because it is driven by gradle
+# add_app(HelloAndroid)  # don't build HelloAndroid here because it is driven by gradle
 # add_app(HelloBaremetal)  # cross-compiles for a bare-metal ARM target; see apps/HelloBaremetal/README.md
 # add_app(HelloiOS)  # don't build HelloiOS here because it isn't universal.
 add_app(HelloPyTorch)
