+++ apps/HelloAndroid/.gitignore
