+++ apps/HelloAndroid/android_smoke_test.sh
