+++ apps/HelloAndroid/app/src/main/cpp/LockedSurface.cpp
