+++ apps/HelloAndroid/app/src/main/cpp/LockedSurface.h
