+++ apps/HelloAndroid/app/src/main/java/com/android/ex/camera2/blocking/BlockingStateCallback.java
