+++ apps/HelloAndroid/app/src/main/java/com/example/helloandroid/AutoFitSurfaceView.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.helloandroidcamera2;
+package com.example.helloandroid;
 
 import android.content.Context;
 import android.util.AttributeSet;
