+++ apps/HelloAndroid/app/src/main/res/layout/activity_camera.xml
@@ -16,4 +16,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#000"
-    tools:context="com.example.helloandroidcamera2.CameraActivity" />
+    tools:context="com.example.helloandroid.CameraActivity" />
