+++ apps/HelloAndroid/app/src/main/res/values/strings.xml
