+++ apps/HelloAndroid/gradle.properties
